requirements.txt
facebook-wda
fastapi[all]